Add files via upload
maam, when i am opening the console to check the mistakes, something new is coming. i dont understand the meaning as it may have been updated. this is just a prototype of the project as im not very sure of how to use arrays as it has been a long time since i used them